uploader/requirements.txt
pika
psycopg2-binary
python-dotenv
typer
pandas